[RoboCIn] Feat: Refactor neural network architectures to improve modularity and flexibility
