+++ agents/methods/architectures/torch/mlp_policy.py
@@ -44,7 +44,7 @@ class MLPPolicy(nn.Module):
         self.back_bone = [
             nn.Linear(hidden_dim, hidden_dim) for _ in range(back_bone_size)
         ]
-        self.back_bone = nn.ModuleList(self.back_bone)
+        self.back_bone = nn.Sequential(*self.back_bone)
         self.output_layer = nn.Linear(hidden_dim, action_dim)
 
         if initializer is not None:
